Decoupled reply service into service and utils.
